refactor: StationConnections의 역할을 Sections로 이동
