+++ src/test/java/subway/domain/section/SectionsTest.java
@@ -3,6 +3,7 @@ package subway.domain.section;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import subway.domain.station.Station;
+import subway.exception.StationNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,6 +12,7 @@ import java.util.Set;
 import static fixtures.SectionFixtures.*;
 import static fixtures.StationFixtures.*;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class SectionsTest {
 
@@ -82,5 +84,30 @@ class SectionsTest {
 
         // then
         assertThat(containingStationNames).isEqualTo(expect);
+    }
+
+    @Test
+    @DisplayName("상행역과 하행역을 순서에 맞게 정렬하여 역 이름 리스트를 반환한다.")
+    void getSortedStationNamesTest() {
+        // given
+        Sections sections = new Sections(List.of(SECTION_강변역_TO_건대역, SECTION_대림역_TO_잠실역, SECTION_잠실역_TO_강변역));
+        List<String> expectSortedStationNames = List.of(STATION_대림역_NAME, STATION_잠실역_NAME, STATION_강변역_NAME, STATION_건대역_NAME);
+
+        // when
+        List<String> sortedStationNames = sections.getSortedStationNames();
+
+        // then
+        assertThat(sortedStationNames).isEqualTo(expectSortedStationNames);
+    }
+
+    @Test
+    @DisplayName("섹션 정보가 없을 때 상행 종점을 찾을 수 없으므로 예외가 발생한다.")
+    void cannotFindUpStationTest() {
+        // given
+        Sections sections = new Sections(List.of());
+        // when, then
+        assertThatThrownBy(() -> sections.getSortedStationNames())
+                .isInstanceOf(StationNotFoundException.class)
+                .hasMessage("상행 종점을 찾을 수 없습니다.");
     }
 }
